+++ CHANGES.md
@@ -7,6 +7,8 @@ CHANGES
 - BiDirSynchronizer
 - Detect conflicts if both targets are modified since last sync.
 - Distinguish whether a resource was added on local or removed on remote.
+- Optionally prompt for username/password.
+- Colored output
 
 0.2.1
 =====